+++ resources/views/products/index.blade.php
@@ -48,6 +48,7 @@
 @endsection
 
 @push('footer-scripts')
+  <script src="//cdn.jsdelivr.net/npm/sweetalert2@11"></script>
   <script>
     
     $(document).ready(function(){
@@ -105,13 +106,31 @@
 
     function deleteItem(id){
       var product_delete_url = "{{ route('products.delete', ':id') }}";
-      $.ajax({
-        type:'DELETE',
-        url:product_delete_url.replace(':id', id),
-        success:function(categories){
-          window.location.replace("{{ route('products.all') }}");
+      Swal.fire({
+        title: '{{ __("Are you sure?") }}',
+        text: "{{ __('You will not be able to revert this!') }}",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: '{{ __("Yes, delete it!") }}',
+        cancelButtonText: '{{ __("Cancel") }}'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          $.ajax({
+            type:'DELETE',
+            url:product_delete_url.replace(':id', id),
+            success:function(){
+              Swal.fire(
+                '{{ __("Deleted!") }}',
+                '{{ __("Your item has been deleted.") }}',
+                'success'
+              )
+              window.location.replace("{{ route('products.all') }}");
+            }
+          });
         }
-      });
+      })
     }
   </script>
 @endpush
